requirements.txt
MetaTrader5
plotly
pandas
Streamlit
Scikit-learn
python-dotenv
atj_algotrading